+++ src/main/java/com/xqz/seckill/domain/SeckillGoods.java
@@ -14,8 +14,8 @@ public class SeckillGoods implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-    private Integer goodsId;
+    private Long id;
+    private Long goodsId;
     private Double price;
     private Integer stock;
     private Timestamp startDate;
@@ -25,19 +25,19 @@ public class SeckillGoods implements Serializable {
         return serialVersionUID;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getGoodsId() {
+    public Long getGoodsId() {
         return goodsId;
     }
 
-    public void setGoodsId(Integer goodsId) {
+    public void setGoodsId(Long goodsId) {
         this.goodsId = goodsId;
     }
